requirements.txt
numpy
Pillow
pynput
pystray
pywin32
pyyaml
watchdog
wxPython